[Refactor] 서비스 이름 기반 헤더 필터로 수정 (X-DPANG-CLIENT-ROLE -> X-DPANG-SERVICE-NAME)
승현이 코드 참고
