+++ src/main/java/com/example/shoppingmallserver/feign/mileage/MileageFeignClient.java
@@ -22,7 +22,7 @@ public interface MileageFeignClient {
      * @return API 호출의 결과를 담은 ResponseEntity 객체 (굳이 DTO로 받을 이유는 없긴 함)
      */
     @PostMapping("/api/mileage")
-    ResponseEntity<SuccessResponse<MileageResponseDto>> createMileage(@RequestHeader("X-DPANG-CLIENT-ID") Long id, @RequestParam Long userId);
+    ResponseEntity<SuccessResponse<MileageResponseDto>> createMileage(@RequestParam Long userId);
 
     /**
      * 사용자의 마일리지를 삭제하는 API를 호출하는 메서드입니다.
@@ -30,7 +30,7 @@ public interface MileageFeignClient {
      * @return API 호출의 결과를 담은 ResponseEntity 객체 (굳이 DTO로 받을 이유는 없지만 오류를 없애기 위함)
      */
     @DeleteMapping("/api/mileage")
-    ResponseEntity<BaseResponse> deleteMileage(@RequestHeader("X-DPANG-CLIENT-ID") Long id, @RequestParam Long userId);
+    ResponseEntity<BaseResponse> deleteMileage(@RequestParam Long userId);
 
 
 }
